added delete alert box

+++ lib/Features/Products/grocery/view/grocery_sub_category_edit_list.dart
@@ -82,8 +82,41 @@ class GrocerySubCategoryEditList extends StatelessWidget {
                       ),
                       GestureDetector(
                         onTap: () {
-                          groceryViewModel.deleteSubCategory(
-                              context, subCategories[index].id);
+                          showDialog(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return AlertDialog(
+                                    backgroundColor: FbColors.backgroundcolor,
+                                    title: Text(
+                                      "Confirm Deletion",
+                                      style: normalFont2(
+                                          fontsize: 20,
+                                          fontweight: FontWeight.w700,
+                                          color: FbColors.black),
+                                    ),
+                                    content: const Text(
+                                        "Are you sure you want to delete this SubCategory?"),
+                                    actions: [
+                                      TextButton(
+                                        onPressed: () {
+                                          Navigator.of(context).pop();
+                                        },
+                                        child: const Text("Cancel"),
+                                      ),
+                                      TextButton(
+                                          onPressed: () {
+
+                                            groceryViewModel.deleteSubCategory(
+                                                context,
+                                                subCategories[index].id);
+                                          },
+                                          child: Text(
+                                            'Delete',
+                                            style: TextStyle(
+                                                color: FbColors.errorcolor),
+                                          ))
+                                    ]);
+                              });
                         },
                         child: const Icon(
                           Icons.delete,
